requirements.txt
streamlit
opencv-python-headless
pillow
matplotlib
imageio